Move ref.value <-> api.expr.Value converters to io.go
Initial commit to test compilation
